skip if build files not exist

+++ Publish/DevOps2/release-bash/71.file-zip.sh
@@ -16,6 +16,11 @@ export APPNAME=xxxxxx
 #----------------------------------------------
 echo "71.file-zip.sh"
 
+if [ ! -d "$basePath/Publish/release/release" ]; then
+    echo '71.file-zip.sh -> skip for no files exist'
+    exit 0
+fi
+
 docker run --rm -i \
 -v $basePath:/root/code \
 serset/filezip bash -c "
